backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
redis
psycopg2-binary
flask-caching
flask-limiter